feat(input): add functionality to accept car names and attempt counts
Implemented the input functionality to accept the names of racing cars,
separated by commas, and the number of attempts for the race. The car
names are stored in an array, and the number of cars is determined
based on the input.

Additionally, a Car class will be created in a future update.